Add TemporaryWorkspace to auto-cleanup workspaces
The TemporaryWorkspace struct wraps a Workspace struct. It implements
the Drop trait to delete the workspace when it goes out of scope. It
also implements the Deref trait for the Workspace type to make a
TemporaryWorkspace transparently behave like a Workspace.

See: https://doc.rust-lang.org/std/ops/trait.Deref.html