Implement partial tree load for 3 way merge
Skips loading shared nodes while reading the 3 trees into memory in the
3 way merge. This works by using the same functions as the checkout and
fast forward merge code, except on 3 trees instead of 2, and the trees
are also manually filtered for their unique nodes afterwards. This also
required adjusting the filtering function, unique_dir_nodes, to fit this
new format.

As well, this commit fixes a bug with the is_modified_from_node util
function wherein calling it on non-existant paths caused a panic. The
function now just returns false, although this may be subject to change
in the future, as that's just the correct behavior in the only spot this
is possible right now.

Applied linters to multiple modules