Skip adding shared dirs for 3 way commit
Implemented logic to skip adding dirs that were shared between the lca,
base, and merge trees in the 3 way merge. Also some optimizations to the
core add logic, including lifting part of the check for merge conflicts
out of process_add_file and avoiding the re-hashing of files with the
same modified times in determine_file_status. This also required an
update to util::fs::is_modified_from_node to check file size as well as
modified times for the unit tests, although that's not yet reflected in
the rest of the code