Filter nodes in push logic
This uses similar logic to what was implemented for fetch. Basically, we
filter out nodes that we know already exist on the remote based on a
starting commit. This drastically reduces the number of nodes we need to
transfer.

In the future, this logic could be combined with the logic used for
fetch (repositories::tree::get_node_hashes_between_commits), but there
are some differences between the two that we would need to reconcile
first.